docker-compose: Improvement to using mtp-common in editable modified
Configuring docker-compose to use local, editable, `mtp-common` has some
issues which are caused by the fact we're installing a new version of
`mtp-common` but then carry on within the same process which may still
point to old version/paths.

This may cause some assets to not be found.

I'm experimenting with the idea of splitting the `python_dependencies`
task from the `serve` task so that the serving of the application happens
after `mtp-common` is fully replaced (and in a new process).

See:
- https://github.com/ministryofjustice/money-to-prisoners-common/pull/420/files
- https://www.edureka.co/community/21344/how-do-i-execute-multiple-commands-using-docker-compose